// packages/demos/antiky/point-light-expo/src/shaders/onboarding.shader.gen.ts
/* Generated by BroMetal. Do not edit — recompile with `npx brometal dev`. */
import type { CompiledShader } from 'brometal';

const onboardingShader: CompiledShader<{ aPosition: 'vec3'; aUv: 'vec2' }, Record<string, never>, { uAtlas: 'sampler2D'; uOpacity: 'float' }> = {
  wgslSrc: `struct BmUniforms {
  uOpacity : f32,
}
@group(0) @binding(0) var<uniform> bm_u : BmUniforms;
@group(0) @binding(1) var uAtlas : texture_2d<f32>;
@group(0) @binding(2) var uAtlas_sampler : sampler;
struct BmVSIn {
  @location(0) aPosition : vec3f,
  @location(1) aUv : vec2f,
}
struct BmVSOut {
  @builtin(position) bm_position : vec4f,
  @location(0) vUv : vec2f,
}
@vertex
fn vs_main(bm_in : BmVSIn) -> BmVSOut {
  var bm_out : BmVSOut;
  bm_out.vUv = bm_in.aUv;
  bm_out.bm_position = vec4f(bm_in.aPosition.x * 0.64, bm_in.aPosition.y * 0.22 - 0.74, 0.0, 1.0);
  bm_out.bm_position.z = (bm_out.bm_position.z + bm_out.bm_position.w) * 0.5;
  return bm_out;
}
@fragment
fn fs_main(bm_in : BmVSOut) -> @location(0) vec4f {
  let sample = textureSample(uAtlas, uAtlas_sampler, bm_in.vUv);
  return vec4f(sample.xyz, sample.w * bm_u.uOpacity);
}
`,
  attributes: { aPosition: 'vec3', aUv: 'vec2' },
  instanceAttributes: {},
  uniforms: { uAtlas: 'sampler2D', uOpacity: 'float' },
  layout: {"attributes":[{"name":"aPosition","type":"vec3","location":0,"size":3,"divisor":0},{"name":"aUv","type":"vec2","location":1,"size":2,"divisor":0}],"uniforms":[{"name":"uAtlas","type":"sampler2D","kind":"1i","size":1,"unit":0,"textureBinding":1,"samplerBinding":2},{"name":"uOpacity","type":"float","kind":"1f","size":1,"offset":0}],"uniformBlockSize":16},

};

export default onboardingShader;